<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add a release note file using reno if the change needs to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [ ] I have added the tests to cover my changes.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the CONTRIBUTING document.
-->

### Summary

Loading instruction schedule map from a backend becomes significant overhead as the qubit size increases. Currently it constructs `Schedule` instance from `CmdDef` entries, and the schedule evaluates timeslots per every `.insert` call. This is inefficient because schedules in the command definition have been already scheduled on the backend. This PR allows user to create a schedule without creating the timeslot.

### Details and comments

Currently the main user interface of Qiskit pulse programming is the pulse builder, in which a schedule is created as a `ScheduleBlock` instance with relative-timing representation with alignment context. The scheduling is done behind the scene at run time with these pulse transform passes.
https://github.com/Qiskit/qiskit-terra/blob/cffbb849344a501be624a75e64dcf757a41a6921/qiskit/pulse/transforms/base_transforms.py#L25-L28
This means user doesn't need to schedule the instructions by oneself, and in principle we can drop most of functionalities to control timeslots on `Schedule` (currently it supports `shift`, `append`, `replace`, `insert`). To create `Schedule`s we only need `insert` method taking an absolute timing t0, that the transform pass may provide. Thus, eventually the schedule can be just a data container to store the list of (`t0, instruction`) pair -- end-user will only deal with `ScheduleBlock` through the builder, rather than directly constructing `Schedules`s by manually calculating instruction timings. Schedule is then only instantiated by drawer, simulator, and qobj converters (if still supported). Validation of schedule timeslot could be delegated to transpiler passes, or even to backend (if backend can receive schedule block), when we completely migrate to from schedule to the pulse gate execution model. 

Note that still there could be many users who use `Schedule` because of [this paper](https://arxiv.org/abs/2004.06755). 

---
This PR does some simplification of `Schedule` as a first step. Here is some performance comparison.
- separate timeslots from `Schedule`
- add validate option to insert method

```python
backend = FakeWashingtonV2()  # no cache
backend.target
``` 

Current main: 3.51 s ± 222 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
This branch: 885 ms ± 36.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)