@mtreinish 2022-05-06
Looking at the import performance of the `test.python.providers.test_fake_backends` module where we were particularly seeing the overhead of pulse schedule creation this makes a big improvement compared to without this PR: https://github.com/Qiskit/qiskit-terra/issues/7914#issuecomment-1116266222

![Screenshot_2022-05-06_16-56-24](https://user-images.githubusercontent.com/2447371/167215569-a56dac89-9369-4103-80ed-d206f2abe2de.png)

With the cProfile overhead we go from ~20 sec to < 10 sec with just this PR. There's still room for improvement before we can close #7914 but this is a big step towards that.