@nkanazawa1989 2022-05-06
Seems like we need to simplify the pulse parameter parser, which is kind of over engineering for current requirement (i.e. this is kind of symbolic equation parser, but "IBM" backend doesn't return such equations -- not sure about other providers).

https://github.com/Qiskit/qiskit-terra/blob/main/qiskit/pulse/parser.py