@mtreinish 2022-05-06
I'm not aware of any providers that have pulse support beside the IBM ones. That doesn't mean that others don't exist, someone could be using it besides IBM I'm just not aware of any. That being said I'm thinking the next step might be looking at getting away from using pulse defaults completely and parsing the payload by handle for the fake providers. That way we can avoid the overhead and not break any potential other users.