@jakelishman 2023-12-21
Naoki: is this still something you're interested in pursuing, or can it be closed as obsoleted by the new pulse work in progress?